Fix Float32/Float64 type mismatch: change var_score to Float64 throughout
The agglovar merge operation calls map_elements on var_score columns,
which produces Python float (64-bit) values. When var_score is typed as
Float32, Polars raises "unexpected value while building Series of type
Float32; found value of type Float64".

Changes:
- schema.py: Change var_score from Float32 to Float64
- call/intra.py: Add return_dtype=pl.Float64 to map_elements(score_model.gap)
- lgsv/variant.py: Add/change return_dtype to pl.Float64 for gap scoring
- lgsv/interval.py: Change return_dtype from pl.Float32 to pl.Float64

+++ src/pav3/lgsv/variant.py
@@ -789,7 +789,7 @@ class InversionVariant(Variant):
                             .filter(~pl.col('is_aligned'))
                             .select(
                                 pl.col('len_qry')
-                                .map_elements(caller_resources.score_model.gap)
+                                .map_elements(caller_resources.score_model.gap, return_dtype=pl.Float64)
                                 .sum()
                             ).item()
                     )
@@ -912,7 +912,7 @@ class ComplexVariant(Variant):
                 self.df_ref_trace
                 .filter(pl.col('type') == 'DEL')
                 .select((pl.col('end') - pl.col('pos')).map_elements(
-                    caller_resources.score_model.gap, return_dtype=pl.Float32
+                    caller_resources.score_model.gap, return_dtype=pl.Float64
                 ).sum())
                 .item()
             )
